moved private, public key check to utils/tools.py
