+++ backOffice/backOffice.py
@@ -19,8 +19,8 @@ from backOffice.thirdLevelWalletManager import ThirdLevelWalletManager
 
 class BackOffice:
     def __init__(self):
-        helper = Helpers()
-        bot_data = helper.read_json_file(file_name='botSetup.json')
+        self.helper = Helpers()
+        bot_data = self.helper.read_json_file(file_name='botSetup.json')
         self.connection = MongoClient(bot_data['database']['connection'], maxPoolSize=20)
         self.as_connection = motor.motor_asyncio.AsyncIOMotorClient(bot_data['database']['connection'])
         self.twitter_details = bot_data["twitter"]
