+++ utils/tools.py
@@ -1,6 +1,8 @@
 import json
 import os
 import sys
+from stellar_sdk import TextMemo, Account, Keypair
+from stellar_sdk.exceptions import MemoInvalidException, Ed25519PublicKeyInvalidError, Ed25519SecretSeedInvalidError
 
 project_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(project_path)
@@ -45,4 +47,31 @@ class Helpers:
             return True
         except FileExistsError as e:
             print(e)
+            return False
+
+    @staticmethod
+    def check_memo(memo):
+        try:
+            TextMemo(text=memo)
+            return True
+        except MemoInvalidException:
+            return False
+
+    @staticmethod
+    def check_public_key(address: str):
+        try:
+            Account(account_id=address, sequence=0)
+            return True
+        except Ed25519PublicKeyInvalidError:
+            return False
+
+    @staticmethod
+    def check_private_key(private_key: str):
+        """
+        Check if private key constructed matches criteria
+        """
+        try:
+            Keypair.from_secret(private_key)
+            return True
+        except Ed25519SecretSeedInvalidError:
             return False
